Start up and UI Changes
Change how I handle the start up sequence. It syncs now only after the Server is created. Does not need to connect to the lockin. This is a response to increased timeouts.

Added a startup status event to communicate the start up status

Completely redid the Run IV Curve Plugin

Tested getting data back from the Lithography Program.

Changed the class data of Plugin Address to Lockin Address and it now reads from class data on startup, allowing this to be configured before we run the plugin.

You can now copy the plugin port by shift + Clicking on the indicator in the plugin.